Move areDatesEqual.js tests to their own file
